+++ tests/test_bridge_domain.rb
@@ -24,9 +24,8 @@ class TestBridgeDomain < CiscoTestCase
   @@cleaned = false # rubocop:disable Style/ClassVars
 
   def cleanup
-    BridgeDomain.bds.each do |_bd, obj|
-      obj.destroy
-    end
+    remove_all_vlans
+    remove_all_bridge_domains
   end
 
   def setup
@@ -43,17 +42,17 @@ class TestBridgeDomain < CiscoTestCase
   def test_single_bd_create_destroy
     bd = BridgeDomain.new('100', true)
     bds = BridgeDomain.bds
-    assert(bds.key?(100), 'Error: failed to create bridge-domain 100')
+    assert(bds.key?('100'), 'Error: failed to create bridge-domain 100')
 
     bd.destroy
     bds = BridgeDomain.bds
-    refute(bds.key?(100), 'Error: failed to destroy bridge-domain 100')
+    refute(bds.key?('100'), 'Error: failed to destroy bridge-domain 100')
   end
 
   def test_bd_create_if_vlan_exists
     vlan = Vlan.new(100)
-    assert_raises(RuntimeError,
-                  'Vlan already exist did not raise RuntimeError') do
+    assert_raises(CliError,
+                  'Vlan already exist did not raise CliError') do
       BridgeDomain.new(100)
     end
     vlan.destroy
@@ -65,7 +64,7 @@ class TestBridgeDomain < CiscoTestCase
     bd = BridgeDomain.new('100-120')
     bds = BridgeDomain.bds
     BridgeDomain.bd_ids_to_array(create).each do |id|
-      assert(bds.key?(id.to_i), 'Error: failed to create bridge-domain ' << id)
+      assert(bds.key?(id.to_s), 'Error: failed to create bridge-domain ' << id)
     end
     bd.destroy
   end
@@ -77,13 +76,13 @@ class TestBridgeDomain < CiscoTestCase
     bd = BridgeDomain.new(create, true)
     bds = BridgeDomain.bds
     bdlist.each do |id|
-      assert(bds.key?(id.to_i), 'Error: failed to create bridge-domain ' << id)
+      assert(bds.key?(id.to_s), 'Error: failed to create bridge-domain ' << id)
     end
 
     bd.destroy
     bds = BridgeDomain.bds
     bdlist.each do |id|
-      refute(bds.key?(id.to_i), 'Error: failed to destroy bridge-domain ' << id)
+      refute(bds.key?(id.to_s), 'Error: failed to destroy bridge-domain ' << id)
     end
   end
 
@@ -92,7 +91,7 @@ class TestBridgeDomain < CiscoTestCase
     bd = BridgeDomain.new(create)
     bds = BridgeDomain.bds
     BridgeDomain.bd_ids_to_array(create).each do |id|
-      assert(bds.key?(id.to_i), 'Error: failed to create bridge-domain ' << id)
+      assert(bds.key?(id.to_s), 'Error: failed to create bridge-domain ' << id)
     end
     bd.destroy
   end
@@ -131,6 +130,7 @@ class TestBridgeDomain < CiscoTestCase
   end
 
   def test_bd_member_vni
+    mt_full_interface?
     bd = BridgeDomain.new(100)
     curr_vni = bd.member_vni.values.join(',')
     assert_equal(bd.default_member_vni, curr_vni,
@@ -151,6 +151,7 @@ class TestBridgeDomain < CiscoTestCase
   end
 
   def test_mapped_bd_member_vni
+    mt_full_interface?
     bd = BridgeDomain.new(100)
     curr_vni = bd.member_vni.values.join(',')
     assert_equal(bd.default_member_vni, curr_vni,
@@ -162,14 +163,15 @@ class TestBridgeDomain < CiscoTestCase
     assert_equal(vni, curr_vni,
                  'Error: Bridge-Domain is mapped to different vnis')
     vni = '6000'
-    assert_raises(RuntimeError,
-                  'Should raise RuntimeError as BD already mapped to vni ') do
+    assert_raises(CliError,
+                  'Should raise CliError as BD already mapped to vni ') do
       bd.member_vni = vni
     end
     bd.destroy
   end
 
   def test_multiple_bd_vni_mapping
+    mt_full_interface?
     bd = BridgeDomain.new('100,110,120')
     curr_vni = bd.member_vni.values.join(',')
     assert_equal(bd.default_member_vni, curr_vni,
@@ -190,6 +192,7 @@ class TestBridgeDomain < CiscoTestCase
   end
 
   def test_member_vni_empty_assign
+    mt_full_interface?
     bd = BridgeDomain.new(100)
     bd.member_vni = ''
     curr_vni = bd.member_vni.values.join(',')
@@ -206,8 +209,8 @@ class TestBridgeDomain < CiscoTestCase
     assert(bd.fabric_control)
     another_bd = BridgeDomain.new(101)
 
-    assert_raises(RuntimeError,
-                  'BD misconfig did not raise RuntimeError') do
+    assert_raises(CliError,
+                  'BD misconfig did not raise CliError') do
       another_bd.fabric_control = true
     end
     bd.destroy
@@ -215,8 +218,8 @@ class TestBridgeDomain < CiscoTestCase
   end
 
   def test_invalid_bd_create
-    assert_raises(RuntimeError,
-                  'BD misconfig did not raise RuntimeError') do
+    assert_raises(CliError,
+                  'BD misconfig did not raise CliError') do
       BridgeDomain.new('90, 5000-5004,100')
     end
   end